style: api 분리와, response, api 이름 정리
