+++ src/api/refreshableApi/index.ts
@@ -2,3 +2,4 @@ export * from './_api';
 export * from './memoApi';
 export * from './tagApi';
 export * from './user';
+export * from './paginationApi';
