+++ src/pages/home/subPages/dashboard/components/MemoSectionList/MemoSection/hook/index.ts
@@ -1,0 +1,1 @@
+export { default as useChildTagMemosManager } from './useChildTagMemosManager';
