+++ src/pages/home/subPages/interfaces/TagRelation/TagRelation.ts
@@ -1,0 +1,6 @@
+import { Tag } from '../MemoInterface';
+
+export interface TagRelation {
+  tag: Tag;
+  child_tags: Tag[];
+}
