+++ src/pages/home/subPages/interfaces/TagRelation/index.ts
@@ -1,0 +1,1 @@
+export * from './TagRelation';
