+++ src/pages/home/subPages/interfaces/index.ts
@@ -1,2 +1,3 @@
 export * from './MemoInterface';
 export * from './MemoSearch';
+export * from './TagRelation';
