+++ src/pages/home/subPages/searchHistory/SearchHistory.tsx
@@ -30,7 +30,7 @@ const SearchHistoryPage = ({}: {}) => {
         searchHistoryLimit: SEARCH_HISTORY_LIMIT,
       });
 
-      if (!Api.isGetSearchHistories(response)) {
+      if (!Api.isSearchHistoriesResponse(response)) {
         throw new Error('메모 검색 기록을 가져오는 중 오류가 발생했습니다.');
       }
       return response;
